requirements.txt
customtkinter==5.2.2
tkintermapview==1.29
pillow==10.4.0
setuptools==75.1.0